added analytics filter types to server-proxy
