+++ cvat-core/src/analytics-report.ts
@@ -2,40 +2,10 @@
 //
 // SPDX-License-Identifier: MIT
 
+import {
+    SerializedAnalyticsEntry, SerializedAnalyticsReport, SerializedDataEntry, SerializedTransformationEntry,
+} from './server-response-types';
 import { ArgumentError } from './exceptions';
-
-export interface SerializedDataEntry {
-    date?: string;
-    value?: number | Record<string, number>
-}
-
-export interface SerializedTransformBinaryOp {
-    left: string;
-    operator: string;
-    right: string;
-}
-
-export interface SerializedTransformationEntry {
-    name: string;
-    binary?: SerializedTransformBinaryOp;
-}
-
-export interface SerializedAnalyticsEntry {
-    name?: string;
-    title?: string;
-    description?: string;
-    granularity?: string;
-    default_view?: string;
-    data_series?: Record<string, SerializedDataEntry[]>;
-    transformations?: SerializedTransformationEntry[];
-}
-
-export interface SerializedAnalyticsReport {
-    id?: number;
-    target?: string;
-    created_date?: string;
-    statistics?: SerializedAnalyticsEntry[];
-}
 
 export enum AnalyticsReportTarget {
     JOB = 'job',
