+++ cvat-core/src/quality-report.ts
@@ -2,35 +2,7 @@
 //
 // SPDX-License-Identifier: MIT
 
-export interface SerializedQualityReportData {
-    id?: number;
-    parent_id?: number;
-    task_id?: number;
-    job_id?: number;
-    target: string;
-    created_date?: string;
-    gt_last_updated?: string;
-    summary?: {
-        frame_count: number,
-        frame_share: number,
-        conflict_count: number,
-        valid_count: number,
-        ds_count: number,
-        gt_count: number,
-        error_count: number,
-        warning_count: number,
-        conflicts_by_type: {
-            extra_annotation: number,
-            missing_annotation: number,
-            mismatching_label: number,
-            low_overlap: number,
-            mismatching_direction: number,
-            mismatching_attributes: number,
-            mismatching_groups: number,
-            covered_annotation: number,
-        }
-    };
-}
+import { SerializedQualityReportData } from './server-response-types';
 
 export interface QualitySummary {
     frameCount: number;
