+++ cvat-core/src/server-response-types.ts
@@ -7,7 +7,7 @@ import {
     DimensionType, JobStage, JobState, JobType, ProjectStatus,
     ShapeType, StorageLocation, LabelType,
     ShareFileType, Source, TaskMode, TaskStatus,
-    CloudStorageCredentialsType, CloudStorageProviderType,
+    CloudStorageCredentialsType, CloudStorageProviderType, ObjectType,
 } from './enums';
 
 export interface SerializedAnnotationImporter {
@@ -24,12 +24,19 @@ export interface SerializedAnnotationFormats {
     importers: SerializedAnnotationImporter[];
     exporters: SerializedAnnotationExporter[];
 }
-export interface ProjectsFilter {
+
+export interface CommonFilterParams {
     page?: number;
-    id?: number;
-    sort?: string;
-    search?: string;
+    page_size?: number | 'all';
     filter?: string;
+    sort?: string;
+    org_id?: number;
+    org?: string;
+    search?: string;
+}
+
+export interface ProjectsFilter extends CommonFilterParams {
+    id?: number;
 }
 
 export interface SerializedUser {
@@ -225,6 +232,10 @@ export interface SerializedOrganization {
     contact?: SerializedOrganizationContact,
 }
 
+export interface QualitySettingsFilter extends CommonFilterParams {
+    task_id?: number;
+}
+
 export interface SerializedQualitySettingsData {
     id?: number;
     task?: number;
@@ -240,6 +251,108 @@ export interface SerializedQualitySettingsData {
     object_visibility_threshold?: number;
     panoptic_comparison?: boolean;
     compare_attributes?: boolean;
+}
+
+export interface QualityConflictsFilter extends CommonFilterParams {
+    report_id?: number;
+}
+
+export interface SerializedAnnotationConflictData {
+    job_id?: number;
+    obj_id?: number;
+    type?: ObjectType;
+    shape_type?: string | null;
+    conflict_type?: string;
+    severity?: string;
+}
+
+export interface SerializedQualityConflictData {
+    id?: number;
+    frame?: number;
+    type?: string;
+    annotation_ids?: SerializedAnnotationConflictData[];
+    data?: string;
+    severity?: string;
+    description?: string;
+}
+
+export interface QualityReportsFilter extends CommonFilterParams {
+    parent_id?: number;
+    peoject_id?: number;
+    task_id?: number;
+    job_id?: number;
+    target?: string;
+}
+
+export interface SerializedQualityReportData {
+    id?: number;
+    parent_id?: number;
+    task_id?: number;
+    job_id?: number;
+    target: string;
+    created_date?: string;
+    gt_last_updated?: string;
+    summary?: {
+        frame_count: number,
+        frame_share: number,
+        conflict_count: number,
+        valid_count: number,
+        ds_count: number,
+        gt_count: number,
+        error_count: number,
+        warning_count: number,
+        conflicts_by_type: {
+            extra_annotation: number,
+            missing_annotation: number,
+            mismatching_label: number,
+            low_overlap: number,
+            mismatching_direction: number,
+            mismatching_attributes: number,
+            mismatching_groups: number,
+            covered_annotation: number,
+        }
+    };
+}
+
+export interface SerializedDataEntry {
+    date?: string;
+    value?: number | Record<string, number>
+}
+
+export interface SerializedTransformBinaryOp {
+    left: string;
+    operator: string;
+    right: string;
+}
+
+export interface SerializedTransformationEntry {
+    name: string;
+    binary?: SerializedTransformBinaryOp;
+}
+
+export interface SerializedAnalyticsEntry {
+    name?: string;
+    title?: string;
+    description?: string;
+    granularity?: string;
+    default_view?: string;
+    data_series?: Record<string, SerializedDataEntry[]>;
+    transformations?: SerializedTransformationEntry[];
+}
+
+export interface AnalyticsReportFilter extends CommonFilterParams {
+    peoject_id?: number;
+    task_id?: number;
+    job_id?: number;
+    start_date?: string;
+    end_date?: string;
+}
+
+export interface SerializedAnalyticsReport {
+    id?: number;
+    target?: string;
+    created_date?: string;
+    statistics?: SerializedAnalyticsEntry[];
 }
 
 export interface SerializedInvitationData {
